Add a single thread for Pygments processing
Pygments (or PythonKit, not sure which) doesn't do well with the multi-threaded piplines Publish uses to parse Markdown files. I've added a global Actor to disable concurrent syntax highlighting, and forced Publish to access this actor sequentially.
